# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog API",
        "description": "API для блога с авторизацией и публикациями",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/auth/register": {
            "post": {
                "summary": "Регистрация нового пользователя",
                "description": "Создает нового пользователя и возвращает токен доступа",
                "operationId": "d764dd091cc4494ae0baf360b03319f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Пользователь успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Пользователь с таким email уже существует",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConflictError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "summary": "Авторизация пользователя",
                "description": "Аутентифицирует пользователя и возвращает токен доступа",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная авторизация",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неверные учетные данные",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "summary": "Выход пользователя",
                "description": "Удаляет токен доступа пользователя",
                "operationId": "69281b12abb272c76871f19cb17ca563",
                "responses": {
                    "200": {
                        "description": "Успешный выход",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Вы успешно вышли из системы"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/auth/me": {
            "get": {
                "summary": "Получение текущего пользователя",
                "description": "Возвращает данные текущего аутентифицированного пользователя",
                "operationId": "b9abb1a7a74670a19c215c2c133f14d8",
                "responses": {
                    "200": {
                        "description": "Данные пользователя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "UserData": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "019dd696-7ce8-73b1-a8b7-28c477e43b7e"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "admin"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
            "AuthResponse": {
                "properties": {
                    "access_token": {
                        "type": "string",
                        "example": "1|AbCdEfGhIjKlMnOpQrStUvWxYz"
                    },
                    "token_type": {
                        "type": "string",
                        "example": "Bearer"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserData"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 400
                    },
                    "title": {
                        "type": "string",
                        "example": "Validation Error"
                    },
                    "detail": {
                        "type": "string",
                        "example": "The given data was invalid."
                    }
                },
                "type": "object"
            },
            "UnauthorizedError": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 401
                    },
                    "title": {
                        "type": "string",
                        "example": "Unauthorized"
                    },
                    "detail": {
                        "type": "string",
                        "example": "Invalid credentials."
                    }
                },
                "type": "object"
            },
            "ConflictError": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 409
                    },
                    "title": {
                        "type": "string",
                        "example": "Conflict"
                    },
                    "detail": {
                        "type": "string",
                        "example": "User already exists."
                    }
                },
                "type": "object"
            },
...
            "UserResponse": {
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/UserData"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Используйте Bearer токен из Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}